pyproject.toml
[project]
name = "psy-faust-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Igor",email = "104066443+OldXrist@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles (==24.1.0)",
    "aiogram (==3.17.0)",
    "aiohappyeyeballs (==2.4.4)",
    "aiohttp (==3.11.11)",
    "aiosignal (==1.3.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "attrs (==24.3.0)",
    "certifi (==2024.12.14)",
    "colorama (==0.4.6)",
    "distro (==1.9.0)",
    "frozenlist (==1.5.0)",
    "future (==1.0.0)",
    "groq (==0.15.0)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jiter (==0.8.2)",
    "magic-filter (==1.0.12)",
    "multidict (==6.1.0)",
    "propcache (==0.2.1)",
    "pydantic (==2.10.5)",
    "pydantic-core (==2.27.2)",
    "pydub (==0.25.1)",
    "python-dotenv (==1.0.1)",
    "sniffio (==1.3.1)",
    "tqdm (==4.67.1)",
    "typing-extensions (==4.12.2)",
    "yarl (==1.18.3)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "exceptiongroup (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
